Oprava testu 66
